+++ WarZeub2/utils/util.h
@@ -26,11 +26,11 @@ enum EDir
 
 // ============================================================================
 
-extern int2 dirs[MAX_DIRS];
+extern float2 dirs[MAX_DIRS];
 
 // ============================================================================
 
-EDir DirectionToTarget(const int2& parSrc, const int2& parDst);
+EDir DirectionToTarget(const float2& parSrc, const float2& parDst);
 
 // ============================================================================
 // ----------------------------------------------------------------------------
@@ -52,10 +52,10 @@ T Clamp(T parValue, T parMin, T parMax)
 
 bool DoesBBoxesCollide(const SDL_Rect* parBoxA, const SDL_Rect* parBoxB);
 
-void TransformToScreenCoordinate(SDL_Rect& parRect, const int2& parCameraPos);
-void TransformToScreenCoordinate(int2& parPos, const int2& parCameraPos);
-void TransformToWorldCoordinate(SDL_Rect& parRect, const int2& parCameraPos);
-void TransformToWorldCoordinate(int2& parRect, const int2& parCameraPos);
+void TransformToScreenCoordinate(SDL_Rect& parRect, const float2& parCameraPos);
+void TransformToScreenCoordinate(float2& parPos, const float2& parCameraPos);
+void TransformToWorldCoordinate(SDL_Rect& parRect, const float2& parCameraPos);
+void TransformToWorldCoordinate(float2& parRect, const float2& parCameraPos);
 
 // ============================================================================
 // ----------------------------------------------------------------------------
